remove lines that have no effect
per @bobchennan in fchollet/keras#4475 "According to the FAQ, after changing the trainable property we need to re-compile the model to take effect. But during the training process the model was not re-compiled so I am not sure during the generator training, the weights for descriminator (sic) changed or not.